# docs/swagger.yaml
basePath: /swagger/box-manage-service
definitions:
  controllers.APIResponse:
    properties:
      code:
        example: 200
        type: integer
      data: {}
      message:
        example: 操作成功
        type: string
      timestamp:
        example: "2025-01-26T12:00:00Z"
        type: string
    type: object
...
  controllers.PaginatedResponse:
    properties:
      code:
        example: 200
        type: integer
      data: {}
      message:
        example: 操作成功
        type: string
      page:
        example: 1
        type: integer
      size:
        example: 10
        type: integer
      timestamp:
        example: "2025-01-26T12:00:00Z"
        type: string
      total:
        example: 100
        type: integer
    type: object
...
  models.VideoFile:
    properties:
      bitrate:
        type: integer
      codec:
        type: string
      converted_path:
        type: string
      created_at:
        description: 创建时间
        example: "2025-01-26T12:00:00Z"
        type: string
      description:
        type: string
      display_name:
        type: string
      duration:
        type: number
      error_msg:
        type: string
      file_size:
        type: integer
      format:
        type: string
      frame_rate:
        type: number
      height:
        type: integer
      id:
        description: 主键ID
        example: 1
        type: integer
      name:
        description: 基本信息
        type: string
      original_file_name:
        description: 存储用户上传时的原始文件名
        type: string
      original_path:
        description: 文件信息
        type: string
      play_url:
        type: string
      process_progress:
        type: integer
      status:
        allOf:
        - $ref: '#/definitions/models.VideoFileStatus'
        description: 处理状态
      stream_app:
        description: ZLMediaKit 相关
        type: string
      stream_id:
        type: string
      updated_at:
        description: 更新时间
        example: "2025-01-26T12:00:00Z"
        type: string
      user_id:
        type: integer
      video_source_id:
        description: 关联信息
        type: integer
      width:
        description: 视频属性
        type: integer
    type: object
  models.VideoFileStatus:
    enum:
    - uploaded
    - converting
    - completed
    - ready
    - failed
    - error
    type: string
    x-enum-comments:
      VideoFileStatusCompleted: 转换完成
      VideoFileStatusConverting: 转换中
      VideoFileStatusError: 错误
      VideoFileStatusFailed: 失败
      VideoFileStatusReady: 就绪
      VideoFileStatusUploaded: 已上传
    x-enum-descriptions:
    - 已上传
    - 转换中
    - 转换完成
    - 就绪
    - 失败
    - 错误
    x-enum-varnames:
    - VideoFileStatusUploaded
    - VideoFileStatusConverting
    - VideoFileStatusCompleted
    - VideoFileStatusReady
    - VideoFileStatusFailed
    - VideoFileStatusError
...
  service.UpdateVideoFileRequest:
    properties:
      description:
        type: string
      display_name:
        type: string
      name:
        type: string
    type: object
...
info:
  contact: {}
  description: AI盒子管理系统后端服务，提供盒子管理、模型管理、任务管理、用户管理等功能
  title: AI盒子管理系统 API
  version: "1.0"
paths:
...
  /api/v1/video/files:
    get:
      consumes:
      - application/json
      description: 分页获取视频文件列表，支持多条件搜索
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: page_size
        type: integer
      - description: 文件状态
        enum:
        - uploading
        - processing
        - ready
        - failed
        in: query
        name: status
        type: string
      - description: 用户ID
        in: query
        name: user_id
        type: integer
      - description: 文件名称搜索关键词
        in: query
        name: name
        type: string
      - description: 文件格式筛选
        enum:
        - mp4
        - avi
        - mov
        - mkv
        in: query
        name: format
        type: string
      - description: 关键词搜索（搜索文件名、描述等）
        in: query
        name: keyword
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/controllers.APIResponse'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/controllers.PaginatedResponse'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/models.VideoFile'
                        type: array
                    type: object
              type: object
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/controllers.APIResponse'
      summary: 获取视频文件列表
      tags:
      - 视频文件管理
  /api/v1/video/files/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定的视频文件
      parameters:
      - description: 视频文件ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "404":
          description: 视频文件不存在
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/controllers.APIResponse'
      summary: 删除视频文件
      tags:
      - 视频文件管理
    get:
      consumes:
      - application/json
      description: 根据ID获取视频文件详细信息
      parameters:
      - description: 视频文件ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/controllers.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.VideoFile'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "404":
          description: 视频文件不存在
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/controllers.APIResponse'
      summary: 获取视频文件详情
      tags:
      - 视频文件管理
    put:
      consumes:
      - application/json
      description: 更新视频文件的基本信息
      parameters:
      - description: 视频文件ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.UpdateVideoFileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            allOf:
            - $ref: '#/definitions/controllers.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.VideoFile'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "404":
          description: 视频文件不存在
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/controllers.APIResponse'
      summary: 更新视频文件信息
      tags:
      - 视频文件管理
  /api/v1/video/files/{id}/convert:
    post:
      consumes:
      - application/json
      description: 将视频文件转换为MP4格式
      parameters:
      - description: 视频文件ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 转换启动成功
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "404":
          description: 视频文件不存在
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/controllers.APIResponse'
      summary: 转换视频文件
      tags:
      - 视频文件管理
  /api/v1/video/files/{id}/download:
    get:
      description: 下载指定的视频文件
      parameters:
      - description: 视频文件ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/octet-stream
      responses:
        "200":
          description: 视频文件
          schema:
            type: file
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "404":
          description: 视频文件不存在
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/controllers.APIResponse'
      summary: 下载视频文件
      tags:
      - 视频文件管理
  /api/v1/video/files/upload:
    post:
      consumes:
      - multipart/form-data
      description: 上传视频文件到系统
      parameters:
      - description: 视频文件
        in: formData
        name: file
        required: true
        type: file
      - description: 文件名称
        in: formData
        name: name
        required: true
        type: string
      - description: 文件描述
        in: formData
        name: description
        type: string
      - description: 用户ID
        in: formData
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 上传成功
          schema:
            allOf:
            - $ref: '#/definitions/controllers.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.VideoFile'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/controllers.APIResponse'
      summary: 上传视频文件
      tags:
      - 视频文件管理
...
swagger: "2.0"